requirements.txt
bs4
pandas
google-generativeai